Added option to disable motors on cancel
Added variable to _CLIENT_VARIABLE to select if motors should be disabled after cancelling a print and added if statement to evaluate variable in CANCEL_PRINT.


Note: I am tired of having to add this on every update, I can't remember a single time I cancel a print and not disable motors. its an optional change.